move get-parent-team into domain-service
