+++ src/domain/pair/__tests__/join-pair.test.ts
@@ -16,7 +16,7 @@ import { IPairRepository } from "../interface/pair-repository"
 // domain service
 import { PairFactory } from "../pair-factory"
 import { GetVacantPairList } from "../get-vacant-pair-list"
-import { GetParentTeam } from "../get-parent-team"
+import { GetParentTeam } from "../domain-service/get-parent-team"
 import { GeneratePairName } from "../domain-service/generate-pair-name"
 
 describe("JoinPair", () => {
