+++ src/domain/pair/domain-service/get-parent-team.ts
@@ -1,7 +1,7 @@
-import { DomainService } from "../shared/domainService"
-import { Pair } from "./pair"
-import { Team } from "../team/team"
-import { ITeamRepository } from "../team/interface/team-repository"
+import { DomainService } from "../../shared/domainService"
+import { Pair } from "../pair"
+import { Team } from "../../team/team"
+import { ITeamRepository } from "../../team/interface/team-repository"
 
 export class GetParentTeam extends DomainService<"get-parent-team"> {
   constructor(private readonly teamRepository: ITeamRepository) {
