+++ src/domain/pair/join-pair.ts
@@ -5,7 +5,7 @@ import { Participant } from "../participant/participant"
 import { IPairRepository } from "./interface/pair-repository"
 import { PairFactory } from "./pair-factory"
 import { GetVacantPairList } from "./get-vacant-pair-list"
-import { GetParentTeam } from "./get-parent-team"
+import { GetParentTeam } from "./domain-service/get-parent-team"
 import { GeneratePairName } from "./domain-service/generate-pair-name"
 
 export class JoinPair extends DomainService<"join-pair"> {
